refactor: 게시글 공통 API를 PostService로 분리 (#45) - PostService 신규 생성: 좋아요, 댓글 관련 공통 API 이관   (/api/posts/... 엔드포인트) - ThemeService에서 공통 메서드 제거 (테마 전용 API만 유지) - usePrefer, useComments, useCommentActions를 theme/ → common/으로 이동   및 PostService 참조로 변경 - CommentModal, ThemeDetailCard, DesignDetailCard import 경로 업데이트 - Detail 페이지(테마/디자인)에서 mock 데이터 제거, 슬라이드 네비게이션 코드 유지
