+++ src/components/community/design/DesignDetailCard.tsx
@@ -4,9 +4,9 @@ import { useNavigate } from "react-router-dom";
 import type { IDesignBoardDetail } from "../../../types/community/design";
 import Text from "../../common/Text";
 import { useOutsideClick } from "../../../services/hooks/common/useOutsideClick";
-import { useComments } from "../../../services/hooks/theme/useComments";
-import { useCommentActions } from "../../../services/hooks/theme/useCommentActions";
-import { usePrefer } from "../../../services/hooks/theme/usePrefer";
+import { useComments } from "../../../services/hooks/common/useComments";
+import { useCommentActions } from "../../../services/hooks/common/useCommentActions";
+import { usePrefer } from "../../../services/hooks/common/usePrefer";
 import { useDeleteDesignPost } from "../../../services/hooks/design/useDeleteDesignPost";
 import { useAuthStore } from "../../../stores/authStore";
 import ImageSlider from "../../common/ImageSlider";
