+++ src/components/community/theme/ThemeDetailCard.tsx
@@ -4,9 +4,9 @@ import { useNavigate } from "react-router-dom";
 import type { IThemeBoardDetail } from "../../../types/community/theme";
 import Text from "../../common/Text";
 import { useOutsideClick } from "../../../services/hooks/common/useOutsideClick";
-import { useComments } from "../../../services/hooks/theme/useComments";
-import { useCommentActions } from "../../../services/hooks/theme/useCommentActions";
-import { usePrefer } from "../../../services/hooks/theme/usePrefer";
+import { useComments } from "../../../services/hooks/common/useComments";
+import { useCommentActions } from "../../../services/hooks/common/useCommentActions";
+import { usePrefer } from "../../../services/hooks/common/usePrefer";
 import { useDeleteThemePost } from "../../../services/hooks/theme/useDeleteThemePost";
 import { useAuthStore } from "../../../stores/authStore";
 import ImageSlider from "../../common/ImageSlider";
