+++ src/services/hooks/common/useComments.ts
@@ -1,11 +1,11 @@
 import { useGetQuery } from '../../api/useApi';
-import { ThemeService } from '../../api/ThemeService';
+import { PostService } from '../../api/PostService';
 import type { ICommentRaw } from '../../../types/community/theme';
 
 export function useComments(postId: number) {
   const { data, isLoading, isError } = useGetQuery<ICommentRaw[]>(
     ['comments', postId],
-    () => ThemeService.getComments(postId, 0, 100),
+    () => PostService.getComments(postId, 0, 100),
     { enabled: !!postId },
   );
